# docs/接口文档/微友App-openapi.yaml
openapi: 3.0.3
info:
  title: 微友 App Backend API Draft
  version: 0.1.0-draft
  description: |
    基于当前后端控制器、需求文档与已生成接口文档整理的 OpenAPI 3.0 草稿。
    
    说明：
    - HTTP Base URL: `http://localhost:8080/api`
    - WebSocket URL: `ws://localhost:8081/ws/chat`
    - WebSocket 协议详情见 `docs/接口文档/微友App-后端接口文档.md`
    - 当前部分接口仍为 demo/stub/no-op 实现，字段结构可用于联调，业务结果未必代表最终生产行为
servers:
  - url: http://localhost:8080/api
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /group/create:
    post:
      tags:
        - Group
      summary: 创建群聊
      description: 当前 groupId 为示例生成值。
      operationId: groupCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupDetailData'
  /group/detail:
    get:
      tags:
        - Group
      summary: 查询群详情
      description: 当前返回 demo 群详情。
      operationId: groupDetail
      parameters:
        - in: query
          name: groupId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupDetailData'
  /group/member/invite:
    post:
      tags:
        - Group
      summary: 邀请成员入群
      description: 会更新群成员数量与成员列表。
      operationId: groupMemberInvite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteMemberRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupDetailData'
  /group/member/list:
    get:
      tags:
        - Group
      summary: 查询群成员列表
      operationId: groupMemberList
      parameters:
        - in: query
          name: groupId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupMemberDataList'
  /group/notice/update:
    post:
      tags:
        - Group
      summary: 更新群公告
      operationId: groupNoticeUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupNoticeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupDetailData'
  /group/setting/update:
    post:
      tags:
        - Group
      summary: 更新群设置
      operationId: groupSettingUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupSettingRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupDetailData'
  /group/admin/update:
    post:
      tags:
        - Group
      summary: 设置管理员
      operationId: groupAdminUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupAdminRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupDetailData'
  /group/member/mute:
    post:
      tags:
        - Group
      summary: 单独禁言成员
      operationId: groupMemberMute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupMemberMuteRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupDetailData'
  /group/member/nickname/update:
    post:
      tags:
        - Group
      summary: 更新群昵称
      operationId: groupMemberNicknameUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupNicknameRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupDetailData'
  /group/member/remove:
    post:
      tags:
        - Group
      summary: 移除群成员
      operationId: groupMemberRemove
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveGroupMemberRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupDetailData'
  /group/owner/transfer:
    post:
      tags:
        - Group
      summary: 转让群主
      operationId: groupOwnerTransfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferGroupOwnerRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupDetailData'
  /group/leave:
    post:
      tags:
        - Group
      summary: 退出群聊
      operationId: groupLeave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveGroupRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupLeaveResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseBase:
      type: object
      required:
        - code
        - message
        - traceId
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        traceId:
          type: string
          example: 3f4b9d6e8a2c4f1b
...
    CreateGroupRequest:
      type: object
      required:
        - groupName
        - memberIds
      properties:
        groupName:
          type: string
        memberIds:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
    InviteMemberRequest:
      type: object
      required:
        - groupId
        - memberIds
      properties:
        groupId:
          type: integer
          format: int64
        memberIds:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
    UpdateGroupNoticeRequest:
      type: object
      required:
        - groupId
        - notice
      properties:
        groupId:
          type: integer
          format: int64
        notice:
          type: string
    UpdateGroupSettingRequest:
      type: object
      required:
        - groupId
      properties:
        groupId:
          type: integer
          format: int64
        muteAll:
          type: boolean
        joinPolicy:
          type: integer
          format: int32
    UpdateGroupAdminRequest:
      type: object
      required:
        - groupId
        - userId
      properties:
        groupId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        admin:
          type: boolean
    UpdateGroupMemberMuteRequest:
      type: object
      required:
        - groupId
        - userId
      properties:
        groupId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        muted:
          type: boolean
    UpdateGroupNicknameRequest:
      type: object
      required:
        - groupId
        - userId
        - groupNickname
      properties:
        groupId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        groupNickname:
          type: string
    RemoveGroupMemberRequest:
      type: object
      required:
        - groupId
        - userId
      properties:
        groupId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
    TransferGroupOwnerRequest:
      type: object
      required:
        - groupId
        - targetUserId
      properties:
        groupId:
          type: integer
          format: int64
        targetUserId:
          type: integer
          format: int64
    LeaveGroupRequest:
      type: object
      required:
        - groupId
      properties:
        groupId:
          type: integer
          format: int64
    GroupDetailData:
      type: object
      properties:
        groupId:
          type: integer
          format: int64
        groupName:
          type: string
        groupAvatar:
          type: string
        ownerUserId:
          type: integer
          format: int64
        memberCount:
          type: integer
          format: int32
        muteAll:
          type: boolean
        joinPolicy:
          type: integer
          format: int32
        notice:
          type: string
    GroupMemberData:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        avatar:
          type: string
        role:
          type: string
        groupNickname:
          type: string
        owner:
          type: boolean
        admin:
          type: boolean
        muted:
          type: boolean
    GroupLeaveResult:
      type: object
      properties:
        groupId:
          type: integer
          format: int64
        left:
          type: boolean
        newOwnerUserId:
          type: integer
          format: int64
          nullable: true
        deleted:
          type: boolean
...
    ApiResponseGroupDetailData:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/GroupDetailData'
    ApiResponseGroupMemberDataList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/GroupMemberData'
    ApiResponseGroupLeaveResult:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/GroupLeaveResult'
...